ENH: Moved property to separate model, redesigned surface-property api
