+++ topobank/manager/admin.py
@@ -1,11 +1,6 @@
 from django.contrib import admin
 
-from .models import Property, Surface, Topography
-
-
-@admin.register(Property)
-class PropertyAdmin(admin.ModelAdmin):
-    list_display = ("id", "name", "value_categorical", "value_numerical", "unit")
+from .models import Surface, Topography
 
 
 @admin.register(Surface)
